// src/app/models/data-from-register-form.model.ts
export interface FinalDataFromRegisterForm {
      tz: string;
      email: string;
      firstname: string;
      lastname: string;
      afterSunset: boolean;
      street: string;
      street2: string;
      city: string;
      state: string;
      country: string;
      zip: string;
      cellphone: string;
      phone: string;
      birthdate: string;
      remark: string;
      groups: { groupId: number }[];
     
}